MOD: 채팅 컨트롤러 패키지 변경 #67
